donationId grow to uint128 for 340282366920938463463374607431768211455 total donation storage

+++ contracts/DonationCore.sol
@@ -31,7 +31,7 @@ contract DonationCore is ERC721BasicToken, DaiDonation {
 
   // Donation struct
   struct Donation {
-    uint32 donationId;  // 4 bytes
+    uint128 donationId;  // 4 bytes
     uint128 amount;     // 16 bytes
     address donor;      // 20 bytes
   }
